Fix exceptions. Add throw wrappings to FilterSerializeException or Parser.ParseException.
