+++ LoC/src/loc/filter/filters/Aggregate/Or.java
@@ -1,7 +1,9 @@
 package loc.filter.filters.Aggregate;
 
+import loc.filter.FilterSerializeException;
 import loc.filter.FilterSerializer;
 import loc.filter.Filter;
+import loc.filter.Parser;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -14,19 +16,19 @@ public class Or extends AggregateFilter {
 
 	public static class Serializer implements FilterSerializer {
 		@Override
-		public Or serialize(String string) throws Exception {
-			String filterSequenceString = new loc.Parser(string)
+		public Or serialize(String string) throws FilterSerializeException {
+			String filterSequenceString = new Parser(string)
 					.skipSpaces()
 					.skipChar(prefix)
 					.skipSpaces()
 					.openParenthesis()
 					.getCurrentBufferString();
-			Filter[] filters = new loc.Parser(filterSequenceString).parseSequence();
+			Filter[] filters = new Parser(filterSequenceString).parseSequence();
 			return new Or(filters);
 		}
 
 		@Override
-		public String serialize(Filter filter) throws Exception {
+		public String serialize(Filter filter) throws FilterSerializeException {
 			return new AggregateFilter.Serializer().serialize(filter);
 		}
 	}
