+++ LoC/src/loc/statistics/StatisticsSerializer.java
@@ -1,6 +1,5 @@
 package loc.statistics;
 
-import loc.FilterSequenceStringBuffer;
 import loc.filter.*;
 
 import java.util.*;
